Calculate the worst case allocation inflation
Instead of using a static formula of:
  (VDEV_RAIDZ_MAXPARITY + 1) * SPA_DVAS_PER_BP == 12

When adding/opening a vdev, set spa->spa_worst_alloc to the worst case
for the worst vdev in the pool.

This will lower the excessive inflation for pools that are simple
spripes or mirror sets.

Is it possible that for dRAID the static formula is not enough?

Sponsored-by: Zededa Inc.
Sponsored-by: Klara Inc.
Signed-off-by: Allan Jude <allan@klarasystems.com>